+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = '2.0.2';
+const VERSION = '2.0.3';
 
 /**
  * Display copyright on the admin dashboard.